requirements.txt
Flask-MySQLdb==1.0.1
Flask-WTF==0.15.1
Flask-Login==0.5.0
passlib==1.7.4
Werkzeug==2.0.1